refactor: unify qt viewers
